refactor: refactor role playing example to use react instead of chain of thought
