+++ examples/ai_society/role_playing_by_react.py
