+++ examples/test/test_ai_society_example.py
@@ -14,7 +14,7 @@
 from mock import patch
 
 import examples.ai_society.role_playing
-import examples.ai_society.role_playing_by_chain_of_thought
+import examples.ai_society.role_playing_by_react
 import examples.function_call.role_playing_with_function
 import examples.open_source_models.role_playing_with_open_source_model
 from camel.types import ModelType
@@ -26,7 +26,7 @@ def test_ai_society_role_playing_example():
                                               chat_turn_limit=2)
 
 
-def test_ai_society_role_playing_by_chain_of_thought_example():
+def test_ai_society_role_playing_by_react():
     with patch('time.sleep', return_value=None):
         examples.ai_society.role_playing_by_chain_of_thought.main(
             ModelType.STUB, chat_turn_limit=2)
